+++ app/managers/file_manager.py
@@ -1,11 +1,14 @@
 """
-This module provides the FileManager class, which offers asynchronous
-methods for performing various file operations such as uploading,
-deleting, writing, reading, encrypting, and decrypting files. It
-leverages aiofiles for efficient asynchronous file I/O operations and
-uses the Fernet encryption system to secure data. The methods are
-designed to handle files and data asynchronously to ensure optimal
-performance in high-concurrency environments.
+This module defines the FileManager class, which provides a suite
+of asynchronous methods for performing file operations, including
+uploading, deleting, writing, reading, copying, encrypting, and
+decrypting files. It utilizes the aiofiles library for non-blocking
+file I/O operations and the Fernet encryption system for secure data
+handling. The methods are designed to work efficiently in asynchronous
+contexts, supporting high-performance and scalable applications.
+The class includes functionality to handle different file types,
+such as images and videos, and ensures optimal performance and
+security for file management tasks.
 """
 
 import aiofiles
@@ -17,6 +20,17 @@ from cryptography.fernet import Fernet
 cfg = get_config()
 cipher_suite = Fernet(cfg.FERNET_KEY)
 
+FILE_COPY_CHUNK_SIZE = 1024 * 8  # 8 KB
+IMAGE_MIMETYPES = [
+    "image/jpeg", "image/png", "image/gif", "image/bmp", "image/tiff",
+    "image/webp", "image/svg+xml", "image/x-icon", "image/heif", "image/heic",
+    "image/jp2", "image/avif", "image/apng", "image/x-tiff",
+    "image/x-cmu-raster", "image/x-portable-anymap", "image/x-portable-bitmap",
+    "image/x-portable-graymap", "image/x-portable-pixmap"]
+VIDEO_MIMETYPES = [
+    "video/mp4", "video/avi", "video/mkv", "video/webm", "video/x-msvideo",
+    "video/x-matroska", "video/quicktime"]
+
 
 class FileManager:
     """
@@ -27,6 +41,26 @@ class FileManager:
     are designed to handle I/O operations efficiently in an asynchronous
     manner to support high-performance applications.
     """
+
+    @staticmethod
+    def is_image(mimetype: str) -> bool:
+        """
+        Determines if the given MIME type is classified as an image type
+        by checking it against a predefined set of image MIME types.
+        The check is case-insensitive and returns True if the MIME type
+        matches any of the image types in the list, otherwise False.
+        """
+        return mimetype.lower() in IMAGE_MIMETYPES
+
+    @staticmethod
+    def is_video(mimetype: str) -> bool:
+        """
+        Determines if the given MIME type is classified as a video type
+        by checking it against a predefined set of video MIME types.
+        The check is case-insensitive and returns True if the MIME type
+        matches any of the video types in the list, otherwise False.
+        """
+        return mimetype.lower() in VIDEO_MIMETYPES
 
     @staticmethod
     @timed
@@ -88,3 +122,24 @@ class FileManager:
         Fernet cipher suite, returning the original data.
         """
         return cipher_suite.decrypt(data)
+
+    @staticmethod
+    @timed
+    async def copy(src_path: str, dst_path: str):
+        """
+        Asynchronously copies the contents of a file from src_path to
+        dst_path in chunks. The method opens the source file for reading
+        in binary mode and the destination file for writing in binary
+        mode. It reads from the source file in chunks and writes those
+        chunks to the destination file until the entire file has been
+        copied. The operation is performed asynchronously to avoid
+        blocking the event loop, and errors such as file not found or
+        permission issues are handled gracefully.
+        """
+        async with aiofiles.open(src_path, mode="rb") as src_context:
+            async with aiofiles.open(dst_path, mode="wb") as dst_context:
+                while True:
+                    chunk = await src_context.read(FILE_COPY_CHUNK_SIZE)
+                    if not chunk:
+                        break
+                    await dst_context.write(chunk)